recruitment-screening-agent/requirements.txt
streamlit==1.28.1
openai==1.3.5
groq==0.4.1
faiss-cpu==1.7.4
PyPDF2==3.0.1
python-dotenv==1.0.0
pydantic==2.5.0
numpy==1.24.3